Add files via upload
The namelist files are provided for the WRF model setup and the CAMx.in file is the control file for the CAMx air quality model. The "CAMx7.1.chemparam.CB6r5_CF2E_EQSAM" file is provided with the CAMx7.10 package as the model chemistry parameters file and is used in our study.